+++ src/Helmholtz/helmkernels_hess.f
@@ -81,7 +81,7 @@ cf2py intent(out) pot
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),pot(nd,nt)
@@ -92,7 +92,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d
       complex *16 zkeye,eye,ztmp
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -173,7 +173,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),pot(nd,nt),grad(nd,3,nt)
@@ -185,7 +185,7 @@ c
       real *8 zdiff(3),dd,d
       complex *16 zkeye,eye,cd,cd1,ztmp
       complex *16 ztmp1,ztmp2,ztmp3
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -276,7 +276,7 @@ cf2py intent(out) pot,grad,hess
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),pot(nd,nt),grad(nd,3,nt)
@@ -290,7 +290,7 @@ c
       complex *16 zkeye,zkeyed,eye,cd,cd1,cd2,ztmp
       complex *16 ztmp1,ztmp2,ztmp3
       complex *16 htmp1,htmp2,htmp3,htmp4,htmp5,htmp6
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -392,7 +392,7 @@ cf2py intent(out) pot
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -404,7 +404,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv
       complex *16 zkeye,eye,cd,cd1,dotprod
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -500,7 +500,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -512,7 +512,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv,dinv2
       complex *16 zkeye,eye,cd,cd2,cd3,cd4,dotprod
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -623,7 +623,7 @@ cf2py intent(out) pot,grad,hess
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -638,7 +638,7 @@ c
       complex *16 zkeye,eye,cd,cd2,cd3,cd4,dotprod
       complex *16 cross12,cross13,cross23,zt0,zf1,zf0
       complex *16 cdcommon,cdcommon2,cdcommon3,cdcommon4
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -768,7 +768,7 @@ c
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -780,7 +780,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv
       complex *16 zkeye,eye,cd,cd1,dotprod
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -881,7 +881,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -893,7 +893,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv,dinv2
       complex *16 zkeye,eye,cd,cd2,cd3,cd4,dotprod
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -1009,7 +1009,7 @@ cf2py intent(out) pot,grad,hess
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer(8) ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),dipvec(nd,3,ns)
@@ -1026,7 +1026,7 @@ c
       complex *16 ztmp1,ztmp2,ztmp3
       complex *16 htmp1,htmp2,htmp3,htmp4,htmp5,htmp6
       complex *16 cdcommon,cdcommon2,cdcommon3,cdcommon4
-      integer i,j,idim
+      integer(8) i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
